+++ storybook-webgpu/src/atmosphere/Vanilla-3DTilesRenderer.tsx
@@ -6,8 +6,8 @@ import {
   UpdateOnChangePlugin
 } from '3d-tiles-renderer/plugins'
 import { AgXToneMapping, PerspectiveCamera, Scene, Vector3 } from 'three'
-import { DRACOLoader } from 'three/examples/jsm/loaders/DRACOLoader.js'
-import { mrt, output, pass, toneMapping } from 'three/tsl'
+import { DRACOLoader } from 'three/addons/loaders/DRACOLoader.js'
+import { context, mrt, output, pass, toneMapping } from 'three/tsl'
 import {
   MeshLambertNodeMaterial,
   PostProcessing,
@@ -21,7 +21,7 @@ import {
 } from '@takram/three-atmosphere'
 import {
   aerialPerspective,
-  AtmosphereContextNode,
+  AtmosphereContext,
   AtmosphereLight,
   AtmosphereLightNode
 } from '@takram/three-atmosphere/webgpu'
@@ -73,8 +73,12 @@ async function init(container: HTMLDivElement): Promise<() => void> {
 
   // The atmosphere context manages resources like LUTs and uniforms shared by
   // multiple nodes:
-  const context = new AtmosphereContextNode()
-  context.camera = camera
+  const atmosphereContext = new AtmosphereContext()
+  atmosphereContext.camera = camera
+  renderer.contextNode = context({
+    ...renderer.contextNode.value,
+    getAtmosphere: () => atmosphereContext
+  })
 
   // Sky background is not necessary as AerialPerspectiveNode renders it:
   const scene = new Scene()
@@ -98,7 +102,7 @@ async function init(container: HTMLDivElement): Promise<() => void> {
   )
   tiles.registerPlugin(
     // Replace non-node materials in every tile:
-    new TileMaterialReplacementPlugin(MeshLambertNodeMaterial)
+    new TileMaterialReplacementPlugin(() => new MeshLambertNodeMaterial())
   )
   tiles.registerPlugin(new TilesFadePlugin())
   scene.add(tiles.group)
@@ -108,7 +112,7 @@ async function init(container: HTMLDivElement): Promise<() => void> {
   renderer.library.addLight(AtmosphereLightNode, AtmosphereLight)
 
   // Create the atmospheric light:
-  const light = new AtmosphereLight(context)
+  const light = new AtmosphereLight()
   scene.add(light)
 
   const controls = new GlobeControls(scene, camera, renderer.domElement)
@@ -143,7 +147,7 @@ async function init(container: HTMLDivElement): Promise<() => void> {
   const depthNode = passNode.getTextureNode('depth')
   const velocityNode = passNode.getTextureNode('velocity')
 
-  const aerialNode = aerialPerspective(context, colorNode, depthNode)
+  const aerialNode = aerialPerspective(colorNode, depthNode)
   const lensFlareNode = lensFlare(aerialNode)
   const toneMappingNode = toneMapping(AgXToneMapping, 5, lensFlareNode)
   const taaNode = temporalAntialias(highpVelocity)(
@@ -175,16 +179,16 @@ async function init(container: HTMLDivElement): Promise<() => void> {
     // the current date and optionally the point of observation:
     const matrixECIToECEF = getECIToECEFRotationMatrix(
       date,
-      context.matrixECIToECEF.value
+      atmosphereContext.matrixECIToECEF.value
     )
     getSunDirectionECI(
       date,
-      context.sunDirectionECEF.value,
+      atmosphereContext.sunDirectionECEF.value,
       observerECEF
     ).applyMatrix4(matrixECIToECEF)
     getMoonDirectionECI(
       date,
-      context.moonDirectionECEF.value,
+      atmosphereContext.moonDirectionECEF.value,
       observerECEF
     ).applyMatrix4(matrixECIToECEF)
 
@@ -214,7 +218,7 @@ async function init(container: HTMLDivElement): Promise<() => void> {
     passNode.dispose()
     controls.dispose()
     tiles.dispose()
-    context.dispose()
+    atmosphereContext.dispose()
     renderer.dispose()
   }
 }